* Fixed issue #16 (https://github.com/Cogniac/cogniac-sdk-csharp/issues/16) * CogniacCSharpSDK.dll is now version 1.0.0.9 * CogUpload.exe is now version 1.0.0.4

+++ CogniacCSharpSDK/Connection.cs
@@ -280,7 +280,7 @@ namespace Cogniac
         (
             string forceSet = null,
             string fileName = null,
-            long mediaTimestamp = 0,
+            double mediaTimestamp = 0,
             bool forceOverwrite = true,
             string[] metaTags = null,
             bool isPublic = false,
